python/gradio_project/legal_help_documentation_msme_webapp/requirements.txt
gradio
google-generativeai
PyMuPDF

